+++ flixopt/components.py
@@ -161,6 +161,8 @@ class LinearConverter(Component):
 
     """
 
+    submodel: LinearConverterModel | None
+
     def __init__(
         self,
         label: str,
@@ -376,6 +378,8 @@ class Storage(Component):
         With flow rates in kW, the charge state is therefore (usually) kWh.
         With flow rates in m3/h, the charge state is therefore in m3.
     """
+
+    submodel: StorageModel | None
 
     def __init__(
         self,
@@ -649,6 +653,8 @@ class Transmission(Component):
         standby power consumption.
 
     """
+
+    submodel: TransmissionModel | None
 
     def __init__(
         self,