+++ flixopt/effects.py
@@ -160,6 +160,8 @@ class Effect(Element):
 
     """
 
+    submodel: EffectModel | None
+
     def __init__(
         self,
         label: str,
@@ -454,12 +456,14 @@ class EffectCollection(ElementContainer[Effect]):
     Handling all Effects
     """
 
+    submodel: EffectCollectionModel | None
+
     def __init__(self, *effects: Effect):
         super().__init__(element_type_name='effects')
         self._standard_effect: Effect | None = None
         self._objective_effect: Effect | None = None
 
-        self.submodel: EffectCollectionModel | None = None
+        self.submodel = None
         self.add_effects(*effects)
 
     def create_model(self, model: FlowSystemModel) -> EffectCollectionModel: