+++ flixopt/flow_system.py
@@ -143,6 +143,8 @@ class FlowSystem(Interface, CompositeContainerMixin[Element]):
           connected_and_transformed automatically when trying to solve a calculation.
     """
 
+    model: FlowSystemModel | None
+
     def __init__(
         self,
         timesteps: pd.DatetimeIndex,