+++ CHANGELOG.md
@@ -74,6 +74,7 @@ If upgrading from v2.x, see the [v3.0.0 release notes](https://github.com/flixOp
 ### 📝 Docs
 
 ### 👷 Development
+- Added type hints for submodel in all Interface classes
 
 ### 🚧 Known Issues